% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelR.R
\name{chunk_object}
\alias{chunk_object}
\title{Break an iterable object into a list of equally sized chunks with sequential elements}
\usage{
chunk_object(object, ncores)
}
\arguments{
\item{object}{Any iterable object.}

\item{ncores}{Integer indicating the number of cores that will be used to process object.}
}
\value{
A list of length ncores with the elements of object divided as evenly as possible across it.
}
\description{
Break an iterable object into a list of equally sized chunks with sequential elements
}
