% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelR.R
\name{parallel_call}
\alias{parallel_call}
\title{Execute a parallel for loop evaluating a call on each element in an object}
\usage{
parallel_call(object, ncores, call, packages = NULL, exported_objects = NULL)
}
\arguments{
\item{object}{Any iterable object.}

\item{ncores}{Integer indicating the number of cores to use.}

\item{call}{A call created with quote() to be executed on each element in object.
Must refer to the element as iterator_element inside the call. Should not use the $ operator.}

\item{packages}{A vector with the names of packages to export to the cluster cores. Default is NULL.}

\item{exported_objects}{A vector with the names of objects to export to the cluster nodes. Default is to try to infer them using all.vars(call).}

\item{combine}{A function or the name of a function used to combine the results of the for loop.}
}
\value{
The results of executing the call on each element of object and combining as specified.
}
\description{
Execute a parallel for loop evaluating a call on each element in an object
}
