% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelR.R
\name{parallel_download}
\alias{parallel_download}
\title{Download a group of files in parallel}
\usage{
parallel_download(urls, directory, ncores, print_progress = FALSE)
}
\arguments{
\item{urls}{A vector of URLs for files to download}

\item{directory}{A path to a directory download files into. Will be created if it doesn't exist.}

\item{ncores}{Integer indicating the number of cores to use.}

\item{print_progress}{A logical value indicating whether to print download progress or not. Default is FALSE.}
}
\value{
The paths to the downloaded files.
}
\description{
Download a group of files in parallel
}
