% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelR.R
\name{parallelize}
\alias{parallelize}
\title{Call a function on each element in an object in parallel}
\usage{
parallelize(
  object,
  ncores,
  parallel_function,
  packages = NULL,
  exported_objects = NULL,
  combine_function = c,
  ...
)
}
\arguments{
\item{object}{Any iterable object.}

\item{ncores}{Integer indicating the number of cores to use.}

\item{parallel_function}{Any function which takes a single argument and returns a result.}

\item{packages}{A vector with the names of packages to export to the cluster cores. Default is NULL.}

\item{exported_objects}{A vector with the names of objects to optionally export to the cluster nodes.}

\item{combine_function}{A function or the name of a function used to combine the results of the for loop. Default is c.}

\item{...}{Optional additional arguments to supply to parallale_function.}
}
\value{
The results of executing parallel_function on each element of object and combining as specified.
}
\description{
Call a function on each element in an object in parallel
}
